apply LSP & OCP principle
