+++ ocp/src/main/java/com/codurance/ocp/Employee.java
@@ -2,25 +2,21 @@ package com.codurance.ocp;
 
 public class Employee {
 
-    private int salary;
-    private int bonus;
-    private EmployeeType type;
+    protected int salary;
 
-    Employee(int salary, int bonus, EmployeeType type) {
+    Employee(int salary) {
         this.salary = salary;
-        this.bonus = bonus;
-        this.type = type;
     }
 
-    public int payAmount() {
-        switch (this.type) {
-            case ENGINEER:
-                return salary;
-            case MANAGER:
-                return salary + bonus;
-            default:
-                return 0;
-        }
-    }
+    // public int payAmount() {
+    //     switch (this.type) {
+    //         case ENGINEER:
+    //             return salary;
+    //         case MANAGER:
+    //             return salary + bonus;
+    //         default:
+    //             return 0;
+    //     }
+    // }
 
 }
