--- ocp/src/main/java/com/codurance/ocp/EmployeeType.java
@@ -1,5 +1,0 @@
-package com.codurance.ocp;
-
-public enum EmployeeType {
-    MANAGER, ENGINEER
-}
