+++ ocp/src/main/java/com/codurance/ocp/Engineer.java
@@ -1,0 +1,13 @@
+package com.codurance.ocp;
+
+public class Engineer extends Employee implements Payable {
+
+    public Engineer(int salary) {
+        super(salary);
+    }
+
+    @Override
+    public int payAmount() {
+        return salary;
+    }
+}
