+++ ocp/src/main/java/com/codurance/ocp/Payable.java
@@ -1,0 +1,5 @@
+package com.codurance.ocp;
+
+public interface Payable {
+    int payAmount();
+}
